Enable standalong custom/CMakeLists.txt configure
For reasons I don't know myself right now, I'd like for the custom
directory to be as "independent" as possible. And for reasons I wrote
down in NOTES.md with this commit, that's not 100% possible. So, at the
very least, there's a humble find_package() call for Qt6 and some other
stuff to at least enable CMake's configuration step to go through.